Filter midnight games
unfortunately PA provide 00:00 as start date when they don't have the start date so we can't do anything with these matches. There's no way to know if the game really starts at midnight or if they just happen to not have the value.

Tonight (1st September 2017) there's about a hundred games playing, of which 90 are without the start date. Filtering these makes sense in that context.